+++ src/components/ProductTile/ProductTilePrice.test.tsx
@@ -3,10 +3,10 @@ import { shallow } from 'enzyme';
 import ProductTilePrice from './ProductTilePrice';
 
 const props = {
-    value: 23423,
+    value: '23423',
     unitExtra: 'uniteExtra',
     unitValue: 'uniteValue',
-    discount: 23423,
+    discount: '23423',
 };
 
 describe('<ProductTilePrice />', () => {