+++ src/components/ProductTile/ProductTilePrice.tsx
@@ -6,34 +6,36 @@ import AnimationValue from './ProductTileValue';
 
 interface IProductTilePriceProps {
     /** value */
-    value?: number;
+    value: string;
     /** unitExtra */
-    unitExtra?: string;
+    unitExtra: string;
     /** unitValue */
-    unitValue?: string;
+    unitValue: string;
     /** discount */
-    discount?: number;
+    discount?: string;
 }
 
 const cn = cnCreate('mfui-product-tile-price');
 class ProductTilePrice extends React.Component<IProductTilePriceProps> {
     static propTypes = {
-        value: PropTypes.number,
+        value: PropTypes.string.isRequired,
         unitExtra: PropTypes.string,
         unitValue: PropTypes.string,
-        discount: PropTypes.number,
+        discount: PropTypes.string,
     };
 
     render() {
         const { value, unitExtra, unitValue, discount } = this.props;
-        const valueElem = <AnimationValue value={`${String(value)} ${unitValue}`} />;
+        const valueElem = <AnimationValue value={(discount || value) + ` ${unitValue}`} />;
         const unitElem = <span className={cn('payment-period')}>{unitExtra}</span>;
 
         return (
             <div className={cn('', { discount: !!discount })}>
                 <div className={cn('discount-condition')}>При покупке новой SIM–карты</div>
                 <span className={cn('old-price')}>
-                    <span className={cn('old-price-value')}>{discount} {unitValue}</span>
+                    <span className={cn('old-price-value')}>
+                        {discount ? `${value} ${unitValue}` : ' '}
+                    </span>
                 </span>
                 <span className={cn('actual-price')}>
                     {valueElem} {unitElem}